% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\name{geom_bar_highstyle}
\alias{geom_bar_highstyle}
\title{Custom geom_bar}
\usage{
geom_bar_highstyle(
  mapping = NULL,
  data = NULL,
  ...,
  fill = "cadetblue",
  width = 0.4
)
}
\description{
Purpose: To include by default a smaller bar width because the default default one is too thick
}
\details{
Use: Exactly like geom_bar.
Defaults will be overriden if specified otherwise in the arguments.
}
