% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\name{geom_density_highstyle}
\alias{geom_density_highstyle}
\title{Custom geom_density}
\usage{
geom_density_highstyle(
  mapping = NULL,
  data = NULL,
  ...,
  linewidth = 0.8,
  color = "cadetblue",
  fill = "cadetblue",
  alpha = 0.6
)
}
\description{
Purpose: To include by default a pre-defined linewidth, fill transparency and colour
}
\details{
Use: Exactly like geom_density.
Defaults will be overriden if specified otherwise in the arguments.
}
