% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\name{geom_density_ridges_highstyle}
\alias{geom_density_ridges_highstyle}
\title{Custom geom_density_ridges}
\usage{
geom_density_ridges_highstyle(
  mapping = NULL,
  data = NULL,
  ...,
  alpha = 0.6,
  linewidth = 0.8
)
}
\description{
Purpose: To include by default a pre-defined linewidth, fill transparency and colour
}
\details{
Use: Exactly like geom_density_ridges.
Defaults will be overriden if specified otherwise in the arguments.
}
