% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_theme.R
\docType{data}
\name{theme_barplot_axes}
\alias{theme_barplot_axes}
\title{Theme add-on for barplots/lollipops}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 6.
}
\usage{
theme_barplot_axes
}
\description{
Theme function that edits both axes so they look nice for a barplot
No x axis and no title for y axis. Also no grid.
}
\details{
Use by adding +theme_barplot_axes to your ggplot object
Better do it after +theme_highstyle() so you don't override it.
}
\keyword{datasets}
