% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_theme.R
\name{theme_highstyle}
\alias{theme_highstyle}
\title{Theme}
\usage{
theme_highstyle(base_size = 12, base_family = "sans")
}
\description{
Theme so all plots are homogeneous in terms of
grids, axes, font sizes, etc.
}
\details{
Use either by adding +theme_highstyle()
to your ggplot object or by theme_set(theme_highstyle()) in your script.
}
