% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_theme.R
\docType{data}
\name{theme_no_x_axis}
\alias{theme_no_x_axis}
\title{Theme add-on to kill x axis}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 4.
}
\usage{
theme_no_x_axis
}
\description{
Theme function that deletes the x axis of a plot
}
\details{
Use by adding +theme_no_x_axis to your ggplot object
Better do it after +theme_highstyle() so you don't override it.
}
\keyword{datasets}
