% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\docType{data}
\name{geom_lollipop_highstyle}
\alias{geom_lollipop_highstyle}
\alias{GeomLollipop}
\title{Custom lollipop plot geometry}
\format{
An object of class \code{GeomLollipop} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomLollipop
}
\description{
Purpose: To avoid manually adding points and segments etc
}
\details{
Use: Exactly like any other ggplot geom.
Just use the x, xend, y, yend you want your segmnts to extend to and your points to be at
Defaults will be overriden if specified otherwise in the arguments.
}
\keyword{datasets}
